+++ Assets/Scripts/Combat/EffectDamageSource.cs
@@ -34,7 +34,7 @@ namespace Game.Combat
             IDamageable damageableObject = collision.GetComponent<IDamageable>();
             if (damageableObject != null)
             {
-                int totalDamage = CalculateTotalDamage();
+                float totalDamage = CalculateTotalDamage();
                 DamageNumberSpawner.Instance.SpawnDamageToEnemyNumber(collision.transform.position, totalDamage);
                 damageableObject.TakeDamage(totalDamage);
             }
@@ -47,7 +47,7 @@ namespace Game.Combat
             }
         }
 
-        private int CalculateTotalDamage()
+        private float CalculateTotalDamage()
         {
             int damage = 0;
             if (weaponClass == WeaponClass.Melee)
